Updates to app.py, codarutils.py, configuration file and environment.yml
Added shorts_minute_filter to the configuration file to tell the program which files to send as source inputs to the merge.  For example, if operating with a 25 MHz system and merging seven 10-min interval files, use input files of 30 minutes past the hour “*30” for hourly merged files and “ *[0,3]0” for half hourly merged files.

app.py
Included argument to main function for configuration filename
Added function to load the configuration file

codarutils.py
Updated add_diagnostic_tables to use integer keys.  This makes the code compatible with changes that were made to hfradarpy.
The expected time for a merged files is calculated to work with sampling for different radar systems.  We expect the merge time to fall behind the source time by a certain number of minutes, which is equal to the number of css files divided by 2 (floor division denoted by // in Python) multiplied by the css interval.
Added code to prevent an error when no source files are found for the merge
Added code to delete a merged file if it is based on a single shorts file
Added code to prevent an error when no output file name is returned from run_LLUVMerger