Fix  auto() for Release 21 software
Prior to SeasSonde Radial Suite Release 21 software, the prefixes of radial metric filenames were RDLv and RDLw.  In R21 the prefixes are RDLx and RDLy.  These are the same prefixes used for radial shorts.  This naming change caused catchup() and auto() options to reprocess all data files instead of new files only.  This update to qccodar fixes that issue.